docs: add formatting + schemas guides; align tax details to conventions
