+++ AGENTS.md
@@ -11,6 +11,8 @@
 - Always use the most specific path alias from `tsconfig.json` for imports
 - Use components located in `src/components/ui`. Import with `@ui` alias
 - Place new, reusable components in `src/components/ui`
+- For amounts, currency, and date formatting, see `docs/ai/formatting.md`
+- For API and schema definitions, see `docs/ai/schemas.md`
 - Stabilize code as follows:
   - **`useCallback`**: Only for functions passed as props to `memo()`-wrapped children or used in Hook dependency arrays. Not needed for handlers on plain HTML elements.
   - **`useMemo`**: Only when (a) the computation is expensive (>1ms), (b) the result is an object/array prop to a `memo()`-wrapped child, or (c) the result is used in a Hook dependency array. Never memoize primitives.
